basic minimum working flask server
